Initialize vachette object with required data

Initialize vachette object with required data

vachette_data(
  obs.data,
  typ.data,
  sim.data = NULL,
  covariates,
  ref.dosenr,
  log.x = FALSE,
  iiv.correction = FALSE,
  error.model = c("proportional", "additive"),
  model.name = NULL,
  mappings = NULL
)

Arguments

obs.data

data.frame; Observed data

typ.data

data.frame; Typical (population) curves

sim.data

data.frame; Simulated (VPC) data

covariates

named character vector; Covariate names with reference values in vachette transformation

ref.dosenr

integer; Dose number to use as the reference dose, corresponding to value in "dosenr" column in input data

log.x

logical; Apply log(x) conversion. Default FALSE

iiv.correction

logical; Apply inter-individual variability correction. Default FALSE

error.model

character; Applied error model, "proportional" or "additive". Default "proportional"

model.name

character; Optional model name for plot output

mappings

named character vector; Optional mappings to be included if column names in input data.frame differ from required column names. See Required Columns section:

Value

vachette_data

Details

If "dosenr" column is missing it will be automatically calculated using the priority of available columns:

"EVID": If available in data, "dosenr" will be calculated using cumsum(EVID==1)

"ADDL"/"II": If "ADDL" and "II" are available in data, "dosenr" will be calculated given additional dose number and interval

"AMT": If only "AMT" column exists in data, "dosenr" will be calculated using cumsum(AMT!=0)

Required columns obs.data

"ID" - Subject ID

"x" - Typically time

"PRED" - Population prediction, required if iiv.correction = TRUE

"IPRED" - Individual prediction, required if iiv.correction = TRUE

"OBS" - DV

"dosenr" - Dose number; unique dose number for ID/time point

Required columns typ.data

"ID" - Subject ID

"x" - Typically time

"PRED" - Population prediction

"dosenr" - Dose number; unique dose number for ID/time point

Required columns sim.data

"ID" - Subject ID

"x" - Typically time

"PRED" - Population prediction, required if iiv.correction = TRUE

"IPRED" - Individual prediction, required if iiv.correction = TRUE

"REP" - Replicate number

Examples

obs <- read.csv(system.file(package = "vachette", "examples", "iv-obs.csv"))
typ <- read.csv(system.file(package = "vachette", "examples", "iv-typ-minmax.csv"))

vd <- vachette_data(
  obs.data = obs,
  typ.data = typ,
  covariates = c(WT = 70),
  mappings = c(OBS = "DV", x = "time"),
  model.name  = "IV"
 ) |>
 apply_transformations()
#> `EVID` column found in obs.data, creating `dosenr` column in data for corresponding ref.dosenr value
#> `EVID` column found in typ.data, creating `dosenr` column in data for corresponding ref.dosenr value
#> 
#> ----------------------------------------------------------
#> --------------------- i.ucov =  1 ----------------------
#>  ucov WT region region.type ref    1 30      1        open  No
#> ----------------------------------------------------------
#> Remove noisy ends of the reference and query curves, if applicable
#> Some obs cannot be associated with query curve
#> No minimum or maximum detected
#> Points kept in y-range of 2.5% to 97.5% only
#> Polynomial order for open end curve fitting = 45
#> Searched for: inflec
#> Returned: NA
#> No inflection point found
#> No minimum or maximum detected
#> Points kept in y-range of 2.5% to 97.5% only
#> Polynomial order for open end curve fitting = 70
#> Searched for: inflec
#> Returned: NA
#> No inflection point found
#> First and last landmark available only
#> Removing 0 observations for i.ucov=1
#> No landmarks detected, x=0 assumed first landmark
#> Optimized last segment x.scaling 0.5163
#> 
#> Reference
#>       x  type            y1   0.1 start 1.960397e+012 102.0   end 2.777013e-08
#> 
#> Query
#>         x  type        y1   0.100 start 19.789242 197.451   end       NA
#> 
#> Zero observations cannot be transformed with proportional error model
#> 6 observations excluded for all i.ucov
#> 
#> ----------------------------------------------------------
#> --------------------- i.ucov =  2 ----------------------
#>  ucov WT region region.type ref    2 70      1        open Yes
#> ----------------------------------------------------------
#> Remove noisy ends of the reference and query curves, if applicable
#> Some obs cannot be associated with reference curve
#> No minimum or maximum detected
#> Points kept in y-range of 2.5% to 97.5% only
#> Polynomial order for open end curve fitting = 45
#> Searched for: inflec
#> Returned: NA
#> No inflection point found
#> No minimum or maximum detected
#> Points kept in y-range of 2.5% to 97.5% only
#> Polynomial order for open end curve fitting = 45
#> Searched for: inflec
#> Returned: NA
#> No inflection point found
#> First and last landmark available only
#> REFERENCE! No need to remove 0 observations for i.ucov=2
#> No landmarks detected, x=0 assumed first landmark
#> Optimized last segment x.scaling 1
#> 
#> Reference
#>       x  type            y1   0.1 start 1.960397e+012 102.0   end 2.777013e-08
#> 
#> Query
#>       x  type            y1   0.1 start 1.960397e+012 102.0   end 2.777013e-08
#> 
#> *** LEFT Extension reference curve by exponential extrapolation ***
#>   >> Applying NO-asymptote reference curve LEFT-side extrapolation
#> 15 observations excluded for all i.ucov